Prep for conversion
In preparation for mana conversion, I changed mana particle to count with floats instead of ints.
General idea: mana can be converted into fractions, but only used in wholes. For movement, the fraction moves as a single unit.

+++ GridGame/Assets/_Scripts/ManaParticle.cs
@@ -12,7 +12,7 @@ public struct ManaParticle
 {
     public ManaType type;
     // public ManaAspects aspect;
-    public int quantity;
+    public float quantity;
     public int particleAge;
     public int particleX;
     public int particleY;
